+++ config.json
@@ -1,7 +1,7 @@
 {
-  "prefix" : "!",
-  "ytAPIkey" : "AIzaSyCsWvBBnOQNiznYRFujzYi6yX7JysMC2x8",
-  "token" : "get rekt",
-  "api_key" : "dc6zaTOxFJmzC",
+  "prefix": "!",
+  "ytAPIkey": "AIzaSyCsWvBBnOQNiznYRFujzYi6yX7JysMC2x8",
+  "token": "get rekt",
+  "api_key": "dc6zaTOxFJmzC",
   "url": "http://api.giphy.com/v1/gifs/search"
 }
