+++ package.json
@@ -1,0 +1,38 @@
+{
+  "name": "dr.jarvis",
+  "version": "1.0.0",
+  "description": "A Discord bot written with discord.js",
+  "main": "src/bot.js",
+  "scripts": {
+    "start": "node src/bot.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Carbowix/Dr.Jarvis.git"
+  },
+  "keywords": [
+    "discord",
+    "bot",
+    "clever",
+    "cool"
+  ],
+  "author": "Carbowix",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/Carbowix/Dr.Jarvis/issues"
+  },
+  "homepage": "https://github.com/Carbowix/Dr.Jarvis#readme",
+  "dependencies": {
+    "chalk": "^1.1.3",
+    "cleverbot-node": "^0.2.4",
+    "discord.js": "^10.0.1",
+    "moment": "^2.17.1",
+    "moment-duration-format": "^1.3.0",
+    "return-deep-diff": "^0.2.7",
+    "superagent": "^3.3.1",
+    "util": "^0.10.3",
+    "weather-js": "^1.0.2",
+    "youtube-search": "^1.0.6",
+    "ytdl-core": "^0.7.21"
+  }
+}
